--- src/ZkFold/Base/Protocol/ARK/Protostar/Oracle.hs
@@ -1,6 +1,0 @@
-module ZkFold.Base.Protocol.ARK.Protostar.Oracle where
-
-class RandomOracle a b where
-    oracle :: a -> b
-
-
