+++ src/ZkFold/Base/Protocol/Protostar/FiatShamir.hs
@@ -2,16 +2,16 @@
 {-# LANGUAGE TypeOperators        #-}
 {-# LANGUAGE UndecidableInstances #-}
 
-module ZkFold.Base.Protocol.ARK.Protostar.FiatShamir where
+module ZkFold.Base.Protocol.Protostar.FiatShamir where
 
 import           Data.ByteString                                 (ByteString)
 import           Prelude                                         hiding (length)
 
 import           ZkFold.Base.Algebra.Basic.Class                 (Bits)
 import           ZkFold.Base.Data.ByteString                     (Binary (..))
-import           ZkFold.Base.Protocol.ARK.Protostar.CommitOpen
-import qualified ZkFold.Base.Protocol.ARK.Protostar.SpecialSound as SpS
-import           ZkFold.Base.Protocol.ARK.Protostar.SpecialSound (SpecialSoundProtocol (..), SpecialSoundTranscript)
+import           ZkFold.Base.Protocol.Protostar.CommitOpen
+import qualified ZkFold.Base.Protocol.Protostar.SpecialSound     as SpS
+import           ZkFold.Base.Protocol.Protostar.SpecialSound     (SpecialSoundProtocol (..), SpecialSoundTranscript)
 import           ZkFold.Base.Protocol.NonInteractiveProof        (NonInteractiveProof (..), ToTranscript (..),
                                                                   challenge)
