+++ src/ZkFold/Base/Protocol/Protostar/Oracle.hs
@@ -1,0 +1,6 @@
+module ZkFold.Base.Protocol.Protostar.Oracle where
+
+class RandomOracle a b where
+    oracle :: a -> b
+
+
