+++ src/ZkFold/Base/Protocol/Protostar/Permutation.hs
@@ -1,4 +1,4 @@
-module ZkFold.Base.Protocol.ARK.Protostar.Permutation where
+module ZkFold.Base.Protocol.Protostar.Permutation where
 
 import           Data.Zip                                        (Zip (..))
 import           Prelude                                         hiding (Num (..), zipWith, (!!), (^))
@@ -8,7 +8,7 @@ import           ZkFold.Base.Algebra.Basic.Number
 import           ZkFold.Base.Algebra.Basic.Permutations          (Permutation, applyPermutation)
 import           ZkFold.Base.Algebra.Polynomials.Multivariate    (var)
 import           ZkFold.Base.Data.Vector                         as V
-import           ZkFold.Base.Protocol.ARK.Protostar.SpecialSound (LMap, SpecialSoundProtocol (..),
+import           ZkFold.Base.Protocol.Protostar.SpecialSound     (LMap, SpecialSoundProtocol (..),
                                                                   SpecialSoundTranscript)
 import           ZkFold.Symbolic.MonadCircuit                    (Arithmetic)
